rename NeoDB to NeoDb to match naming convention...update test to check for copyright
